LC-minimum-time-to-complete-trips [Runtime: 51 ms; Memory: 98.3 MB]

+++ array/minimum-time-to-complete-trips.cpp
@@ -12,7 +12,7 @@ public:
         long long right = (long long) *min_element(time.begin(), time.end()) * totalTrips;
 
         while(left <= right){
-            int mid = left + (right - left) / 2;
+            long long mid = left + (right - left) / 2;
             long long trips = 0;
 
             for(int t : time){
